pull update from upstream

+++ src/kem/ml_kem/pqcrystals-kyber-standard_ml-kem-768_ref/poly.c
@@ -19,7 +19,7 @@
 void poly_compress(uint8_t r[KYBER_POLYCOMPRESSEDBYTES], const poly *a)
 {
   unsigned int i,j;
-  int32_t u;
+  int16_t u;
   uint32_t d0;
   uint8_t t[8];
